add do loop with integer and cancel verification
A do loop was added to prompt the user to re-enter
an input if they enter a non-integer value. It
will exit if the user clicks "cancel" on the
prompt window. I did it this way because it is
currently the only way I know how to do it.
Next change plan: verify if user inputs a positive
integer greater than 1.